Reputation becomes an Int, add queries Reputation into an Int. We will need to remove out `double`s and 'float's from .proto files eventually, this is just a start. Queries for reputation/weight within an Epoch, and for the minim validation average, based on current/previous traffic amount. Useful in general, but needed to have a chance to test this
Good chunk of test fixes as well

+++ inference-chain/x/inference/epochgroup/epoch_group.go
@@ -70,7 +70,7 @@ func (eg *EpochGroup) CreateGroup(ctx context.Context) error {
 	return nil
 }
 
-func (eg *EpochGroup) AddMember(ctx context.Context, address string, weight uint64, pubkey string, seedSignature string, reputation float64) error {
+func (eg *EpochGroup) AddMember(ctx context.Context, address string, weight uint64, pubkey string, seedSignature string, reputation int64) error {
 	eg.Logger.LogInfo("Adding member", types.EpochGroup, "address", address, "weight", weight, "pubkey", pubkey, "seedSignature", seedSignature)
 	val, found := eg.GroupDataKeeper.GetEpochGroupData(ctx, eg.GroupData.PocStartBlockHeight)
 	if !found {
@@ -88,7 +88,7 @@ func (eg *EpochGroup) AddMember(ctx context.Context, address string, weight uint
 	eg.GroupData.ValidationWeights = append(eg.GroupData.ValidationWeights, &types.ValidationWeight{
 		MemberAddress: address,
 		Weight:        int64(weight),
-		Reputation:    reputation,
+		Reputation:    int32(reputation),
 	})
 	eg.GroupData.TotalWeight += weight
 	eg.GroupDataKeeper.SetEpochGroupData(ctx, *eg.GroupData)
